move metadata formatting methods to parse
